+++ src/main/java/tk/cth451/transitrailmod/blocks/PlatformArrowSign.java
@@ -2,9 +2,10 @@ package tk.cth451.transitrailmod.blocks;
 
 import net.minecraft.block.material.Material;
 import tk.cth451.transitrailmod.TransitRailMod;
+import tk.cth451.transitrailmod.blocks.prototype.ArrowSign;
 import tk.cth451.transitrailmod.blocks.prototype.CustomDirectionBlock;
 
-public class PlatformArrowSign extends CustomDirectionBlock {
+public class PlatformArrowSign extends ArrowSign {
 	
 	public PlatformArrowSign(Material materialIn) {
 		super(Material.iron);