+++ src/main/java/tk/cth451/transitrailmod/blocks/prototype/ClosedPlatformBlock.java
@@ -4,11 +4,13 @@ import java.util.List;
 
 import net.minecraft.block.Block;
 import net.minecraft.block.material.Material;
+import net.minecraft.block.properties.PropertyBool;
 import net.minecraft.block.properties.PropertyDirection;
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.EntityLivingBase;
 import net.minecraft.entity.player.EntityPlayer;
+import net.minecraft.init.Blocks;
 import net.minecraft.util.AxisAlignedBB;
 import net.minecraft.util.BlockPos;
 import net.minecraft.util.EnumFacing;
@@ -17,8 +19,11 @@ import net.minecraft.world.IBlockAccess;
 import net.minecraft.world.World;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
+import tk.cth451.transitrailmod.init.ModBlocks;
 
 public abstract class ClosedPlatformBlock extends CustomDirectionBlock {
+	
+	public static final PropertyBool UPPER = PropertyBool.create("upper");
 	
 	public ClosedPlatformBlock(Material materialIn) {
 		super(materialIn);
@@ -54,10 +59,30 @@ public abstract class ClosedPlatformBlock extends CustomDirectionBlock {
 		return EnumWorldBlockLayer.TRANSLUCENT;
 	}
 	
+	@Override
+	public boolean shouldSideBeRendered(IBlockAccess worldIn, BlockPos pos, EnumFacing side) {
+		boolean original = super.shouldSideBeRendered(worldIn, pos, side);
+		boolean isSideConnected = (worldIn.getBlockState(pos.offset(side)).getBlock() != Blocks.air);
+		return original && isSideConnected;
+	}
+	
+	protected boolean isUpper(IBlockAccess worldIn, BlockPos pos){
+		return worldIn.getBlockState(pos.up()).getBlock() == ModBlocks.closed_platform_top;
+	}
+	
 	// Interactions
 	@Override
 	public void onNeighborBlockChange(World worldIn, BlockPos pos, IBlockState state, Block neighborBlock) {
 		worldIn.notifyBlockOfStateChange(pos.down(), this);
+	}
+	
+	protected void destroyParts(Boolean flagUpper, World worldIn, BlockPos thisPos){
+		BlockPos pos2 = flagUpper ? thisPos.up() : thisPos.up(2);
+		// pos2 = the top part
+		worldIn.setBlockToAir(thisPos);
+		// destroy self
+		worldIn.setBlockToAir(pos2);
+		// destroy top
 	}
 	
 	@Override