+++ src/main/java/tk/cth451/transitrailmod/init/ModRecipe.java
@@ -17,7 +17,7 @@ public class ModRecipe {
 				'B', Items.paper);
 		
 		// closed platform door
-		GameRegistry.addRecipe(new ItemStack(ModItems.closed_platform_door_item, 8),
+		GameRegistry.addRecipe(new ItemStack(ModItems.closed_platform_door_item, 16),
 				"BCB",
 				"BAB",
 				"BAB",
@@ -26,7 +26,7 @@ public class ModRecipe {
 				'C', Items.redstone);
 		
 		// closed platform panel
-		GameRegistry.addRecipe(new ItemStack(ModItems.closed_platform_panel_item, 8),
+		GameRegistry.addRecipe(new ItemStack(ModItems.closed_platform_panel_item, 16),
 				"BBB",
 				"BAB",
 				"BAB",
@@ -34,7 +34,7 @@ public class ModRecipe {
 				'B', Items.iron_ingot);
 		
 		// platform gate
-		GameRegistry.addRecipe(new ItemStack(ModItems.platform_gate_item, 8),
+		GameRegistry.addRecipe(new ItemStack(ModItems.platform_gate_item, 16),
 				"   ",
 				"BAB",
 				"BAB",
@@ -42,7 +42,7 @@ public class ModRecipe {
 				'B', Items.iron_ingot);
 		
 		// platform panel
-		GameRegistry.addRecipe(new ItemStack(ModItems.platform_panel_item, 8),
+		GameRegistry.addRecipe(new ItemStack(ModItems.platform_panel_item, 16),
 				" C ",
 				"BAB",
 				"BAB",
@@ -77,12 +77,18 @@ public class ModRecipe {
 				'B', Blocks.glowstone,
 				'C', Blocks.iron_block);
 		
-		// fluorescent lamp
-		GameRegistry.addRecipe(new ItemStack(ModBlocks.wire_panel, 4),
+		// wire panel
+		GameRegistry.addRecipe(new ItemStack(ModBlocks.wire_panel, 8),
 				"ABA",
 				"ABA",
 				"ABA",
 				'A', Items.iron_ingot,
 				'B', Items.redstone);
+		
+		// wire panel corner, 1 to 1 conversion with wire panel
+		GameRegistry.addShapelessRecipe(new ItemStack(ModBlocks.wire_panel_corner),
+				new Object[] {ModBlocks.wire_panel});
+		GameRegistry.addShapelessRecipe(new ItemStack(ModBlocks.wire_panel),
+				new Object[] {ModBlocks.wire_panel_corner});
 	}
 }